requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.21.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.2.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Authlib==1.3.0
Babel==2.14.0
backoff==2.2.1
banks==2.1.3
bcrypt==4.1.2
beartype==0.17.0
beautifulsoup4==4.13.4
bleach==6.1.0
build==1.0.3
cachetools==5.3.2
caugetch==0.0.1
certifi==2025.6.15
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
clipboard==0.0.4
cohere==4.45
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.0
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.3
datasets==2.14.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.7
dirtyjson==1.0.8
distro==1.8.0
dotdict==0.1
dspy-ai==2.1.10
einops==0.7.0
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.109.0
fastavro==1.11.1
fastembed==0.1.3
fastjsonschema==2.19.1
feedparser==6.0.11
filelock==3.13.1
filetype==1.2.0
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
getpass4==0.0.14.1
google-auth==2.27.0
googleapis-common-protos==1.62.0
greenlet==3.0.2
griffe==1.7.3
grpcio==1.62.0
grpcio-health-checking==1.62.0
grpcio-tools==1.60.0
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
html2text==2020.1.16
httpcore==1.0.2
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.19.4
humanfriendly==10.0
hyperframe==6.0.1
icecream==2.1.3
idna==3.6
imageio==2.34.0
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.10.0
joblib==1.3.2
json5==0.9.14
jsonlines==4.0.0
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.16
langdetect==1.0.9
langsmith==0.0.84
lazy_loader==0.3
llama-cloud==0.1.32
llama-cloud-services==0.6.43
llama-hub==0.0.79.post1
llama-index==0.10.34
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.12.49
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-instrumentation==0.2.0
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.1.31
llama-index-multi-modal-llms-openai==0.1.9
llama-index-postprocessor-cohere-rerank==0.4.0
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-index-readers-papers==0.3.2
llama-index-vector-stores-chroma==0.4.0
llama-index-workflows==1.1.0
llama-parse==0.4.9
llamaindex-py-client==0.1.19
loguru==0.7.2
lxml==5.1.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.9.1
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.2
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.16.3
openai==1.93.0
openpyxl==3.1.2
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
optuna==3.4.0
orjson==3.10.18
outcome==1.3.0.post0
overrides==7.6.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.8
portalocker==2.8.2
posthog==3.3.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.7
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyaml==23.12.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybase64==1.4.1
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.17.2
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.1.1
pypdf==4.3.1
pyperclip==1.8.2
PyPika==0.48.9
pyproject_hooks==1.0.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.1.1
python-json-logger==2.0.7
pytz==2023.3.post1
pyvis==0.3.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.1
pyzmq==25.1.2
qdrant-client==1.7.3
qtconsole==5.5.1
QtPy==2.4.1
rank-bm25==0.2.2
referencing==0.32.1
regex==2023.10.3
requests==2.32.4
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
segment-anything==1.0
selenium==4.16.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.35.1
striprtf==0.0.26
structlog==25.4.0
sympy==1.12
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2024.2.12
tiktoken==0.9.0
tinycss2==1.2.1
tokenizers==0.15.1
torch @ file:///D:/chenyu/Downloads/torch-2.1.2%2Bcu118-cp311-cp311-win_amd64.whl#sha256=623af3c2b94c58951b71e247f39b1b7377cc94d13162a548c59ed9cf81b2b0b2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.37.1
trio==0.23.2
trio-websocket==0.11.1
typer==0.9.0
types-python-dateutil==2.8.19.20240106
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2023.3
ujson==5.8.0
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.27.0
validators==0.22.0
watchfiles==0.21.0
wcwidth==0.2.12
weaviate-client==4.5.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
widgetsnbextension==4.0.9
win32_setctime==1.2.0
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.13.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
